Improve authorization handling and return codes
We still had a few issues if people tried to push into the root
of the project storage or their users' storage.

Also we often returned not very proper HTTP return codes, which
made it not always that obvious what was going wrong. This means
that we now return a proper 403 if access is denied.

If ro-access is allowed but git repository does not exists we return
a proper 404.